\docType{package}
\name{ALA4R}
\alias{ALA4R}
\alias{ALA4R-package}
\title{\pkg{ALA4R}}
\description{
This project enables the R community to access data and
tools hosted at the Atlas of Living Australia (ALA:
http://www.ala.org.au). The goal is to enable outputs (e.g.
observations of species) to be queried and output in
standard formats for R. ALA4R is based around use of the
extensive web services provided by the Atlas; see
References.
}
\author{
Atlas of Living Australia \email{support@ala.org.au}, Ben
Raymond, Jeremy VanDerWal, Lee Belbin
}
\references{
\url{http://api.ala.org.au/ws} Currently, not all services
of the following pages had been moved to api.ala.org.au The
following web service reference pages are therefore added
for completeness.

\url{http://bie.ala.org.au/ws}

\url{http://biocache.la.org.au/ws}

\url{http://spatial.ala.org.au/ws}
}

