% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_categories.R
\name{ala_categories}
\alias{ala_categories}
\title{List valid options for a categorical field}
\usage{
ala_categories(field, limit = 20)
}
\arguments{
\item{field}{string: field to return the categories for. Use `ala_fields`
to view valid fields.}

\item{limit}{numeric: maximum number of categories to return. 20 by default.}
}
\value{
a dataframe of field name and category name.
}
\description{
Used for checking filter values for `ala_occurrences` and `ala_counts` are
valid
}
\examples{
ala_categories("basis_of_record")
ala_categories("state")
}
