% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_checklist.R
\name{ala_checklist}
\alias{ala_checklist}
\title{List species matching criteria}
\usage{
ala_checklist(taxon_id, filters, area, caching = "off")
}
\arguments{
\item{taxon_id}{string: single species ID or vector of species ids. Use 
`ala_taxa()` to get lookup species id.}

\item{filters}{string: a list to narrow down the search, in the form
`list(field = value)`. To limit the query to a year range, use
`list(year = c(1990, 2000))`. This will return records between and including
the years provided.}

\item{area}{string or sf object: restrict the search to an area. Can provide
sf object, or a wkt string. WKT strings longer than 10000 characters will
not be accepted by the ALA- see the vignette for how to work around this.}
}
\value{
dataframe of matching species, with optional additional details.
}
\description{
List species matching criteria
}
