% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_citation.R
\name{ala_citation}
\alias{ala_citation}
\title{Generate a citation for occurrence data}
\usage{
ala_citation(data)
}
\arguments{
\item{data}{data.frame: occurrence data generated by `ala_occurrences`}
}
\value{
string citation
}
\description{
Generate a citation for occurrence data
}
