% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_columns.R
\name{ala_columns}
\alias{ala_columns}
\title{Build dataframe of columns to keep}
\usage{
ala_columns(group, extra)
}
\arguments{
\item{group}{string: name of column group to include}

\item{extra}{string: additional column names to return}
}
\description{
Build dataframe of columns to keep
}
