% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_config.R
\name{ala_constants}
\alias{ala_constants}
\title{ALA4R internal settings and parameters}
\usage{
ala_constants()
}
\value{
A named list of values.
}
\description{
Most users will not need to use this function: generally, users wishing to adjust ALA4R behaviour should look 
at \code{\link{ala_config}}. However, developers of packages that import/depend on ALA4R might choose to override
this function with their own.
}
\examples{
\dontrun{
 ala_constants()
}
}
\seealso{
\code{\link{ala_config}}
}

