% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_geometry.R
\name{ala_geometry}
\alias{ala_geometry}
\title{Build a WKT string from a spatial object or verify an existing string.}
\usage{
ala_geometry(wkt, area)
}
\arguments{
\item{wkt}{string: wkt to be verified. WKT strings longer than 10000
characters will not be accepted by the ALA- see the vignette for how to
work around this.}

\item{area}{sf object: area to be converted to wkt}
}
\value{
WKT string representing area provided
}
\description{
Used by `ala_occurrences() to restrict search to an area.`
}
