% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_media.R
\name{ala_media}
\alias{ala_media}
\title{Download images and sounds from the ALA.}
\usage{
ala_media(identifier, identifier_type)
}
\arguments{
\item{identifier}{string: a single or vector of identifiers, of type
specified by `identifier_type`}

\item{identifier_type}{string: one of `c("occurrence", "image")`}
}
\description{
Download images and sounds from the ALA.
}
