\name{ala_user_agent}
\alias{ala_user_agent}
\title{User agent string for ALA web requests}
\description{Returns the user-agent string used to make requests from
  the ALA web servers.
}
\usage{
ala_user_agent()
}
%- maybe also 'usage' for other objects documented here.
\details{
  TODO: provide instructions for the user to override this function with
  one that provides more informative details to the ALA, should they
  wish to do so
}
\value{\item{x}{character string, by default "ALA4R [version] (R version/platform)"}
}
\references{
}
\author{Ben Raymond \email{ben@theraymonds.org}, Jeremy VanDerWal \email{jjvanderwal@gmail.com}}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
ala_user_agent()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
