\docType{data}
\name{aus}
\alias{aus}
\title{Basic Australia raster for visualization}
\format{SpatialGridDataFrame (package sp)}
\usage{
data(aus)
}
\description{
A basic raster dataset in SpatialGridDataFrame (package sp)
representing Australia at 0.05 degree (~5km) resolution.
}
\keyword{datasets}

