% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_filters.R
\name{exclude}
\alias{exclude}
\title{Negate a filter value}
\usage{
exclude(value)
}
\arguments{
\item{value}{string: filter value(s) to be excluded}
}
\value{
value with class "exclude"
}
\description{
Negate a filter value
}
