% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/image_info.R
\name{image_info}
\alias{image_info}
\title{Fetch information about an image, given its image ID}
\usage{
image_info(id, verbose = ala_config()$verbose)
}
\arguments{
\item{id}{character: IDs of images (e.g. as returned by \code{\link{occurrences}}  in the imageUrl column). Each ID will be of a format something like "84654e14-dc35-4486-9e7c-40eb2f8d3faa"}

\item{verbose}{logical: show additional progress information? [default is set by ala_config()]}
}
\value{
A data.frame with one row per \code{id}
}
\description{
Note that there is currently no web service that provides image information, and so we are scraping results from pages of the form http://images.ala.org.au/image/details?imageId=id. This web scraping may be fragile, and will be replaced by a web-service-based function when one becomes available.
}
\examples{
image_info(c("84654e14-dc35-4486-9e7c-40eb2f8d3faa","b8344134-254d-4116-a98d-4a37e7362a4e"))
}
\author{
Atlas of Living Australia \email{support@ala.org.au}
}
\seealso{
\code{\link{ala_config}} \code{\link{occurrences}}
}

