% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_search.R
\name{image_search}
\alias{image_search}
\title{Query images based on search terms}
\usage{
image_search(q, fq, download = FALSE, download_path, sounds = FALSE,
  verbose = ala_config()$verbose)
}
\arguments{
\item{q}{string: (optional) free text search query}

\item{fq}{string: (optional) character string or vector of strings,
specifying filters to be applied to the original query. These are of the
form "INDEXEDFIELD:VALUE" e.g. "kingdom:Fungi".}

\item{download}{logical: if TRUE download all images and add location to
dataframe}

\item{download_path}{string: (optional) filepath to download images to.
If not given and download param is TRUE, will create an images
folder}

\item{sounds}{logical (optional) Image search also returns sound files.
Ignored unless explicitly requested.}

\item{verbose}{logical: show additional progress information?
[default is set by ala_config()]}
}
\value{
Data frame of image results
}
\description{
Query images based on search terms
}
\examples{
\dontrun{
## Download all kangaroo images with a CC BY-NC 4.0 licence
image_search(q="kangaroo",fq="recognisedLicence:CC BY-NC 4.0",
download=TRUE)
}
## Download all frog sounds
image_search(q="frog",fq="fileType:sound",download=TRUE, sounds=TRUE)
}
\references{
\itemize{
\item Associated ALA web service for image search counts:
\url{https://images.ala.org.au/ws#/Search/search}
}
}
