% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{images}
\alias{images}
\title{Retrieve images uses image ids}
\usage{
images(id, fq, download = FALSE, download_path, sounds = FALSE,
  verbose = ala_config()$verbose)
}
\arguments{
\item{id}{character: IDs of images to be downloaded as single string or
vector of strings}

\item{fq}{string: (optional) character string or vector of strings,
specifying filters to be applied to the original query. These are of the
form "INDEXEDFIELD:VALUE" e.g. "kingdom:Fungi".}

\item{download}{logical: if TRUE download all images and add location to
dataframe}

\item{download_path}{string: (optional) filepath to download images to.
If not given and download param is TRUE, will create an images
folder}

\item{sounds}{logical (optional) Image search also returns sound files.
Ignored unless explicitly requested.}

\item{verbose}{logical: show additional progress information?
[default is set by ala_config()]}
}
\value{
Data frame of image results
}
\description{
Retrieve images uses image ids
}
\examples{
\dontrun{
## Download all images with a CC BY-NC 4.0 licence
images(id="da5fe120-e213-4cd6-9c5f-62346ed2e466",
fq="recognisedLicence:CC BY-NC-SA 4.0", download=TRUE)
}
}
\references{
\itemize{
\item Associated ALA web service for image retrieval
\url{https://images.ala.org.au/ws#/Search/search}
}
}
