% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occurrence_details.R
\name{occurrence_details}
\alias{occurrence_details}
\title{Retrieve the full details of occurrence records}
\usage{
occurrence_details(uuid, verbose = ala_config()$verbose)
}
\arguments{
\item{uuid}{string: one or more record ids, as returned by \code{\link{occurrences}} (in the \code{data$id} column)}

\item{verbose}{logical: show additional progress information? [default is set by ala_config()]}
}
\value{
A named list (named by uuid), each element of which is a list containing the details for that uuid. This inner list will be empty if no match is found for the supplied uuid
}
\description{
Retrieve the full details of occurrence records
}
\examples{
s1=occurrence_details("f259c5ce-200c-41a2-b73a-e36a91f748f7")
str(s1)
}
\author{
Atlas of Living Australia \email{support@ala.org.au}
}
\references{
Associated ALA web service: \url{http://api.ala.org.au/#ws102}
}
\seealso{
\code{\link{occurrences}} \code{\link{ala_config}}
}

