% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_facets.R
\name{occurrence_facets}
\alias{occurrence_facets}
\title{Get facet counts for occurrence data}
\usage{
occurrence_facets(
  facet,
  query,
  start,
  page_size,
  verbose = ala_config()$verbose
)
}
\arguments{
\item{facet}{string: facet to be used. Currently only one facet is supported}

\item{query}{string: query of the form field:value e.g. q=genus:Macropus or 
a free text search e.g. q=Macropus}

\item{start}{numeric: (optional) (positive integer) start offset for the results}

\item{page_size}{numeric: (optional) (positive integer) maximum number of records 
to return. Defaults to the server-side value - currently 30}

\item{verbose}{logical: show additional progress information? 
[default is set by ala_config()]}
}
\description{
Get facet counts for occurrence data
}
\references{
Associated ALA web service for facet counts \itemize{
\item \url{https://api.ala.org.au/#ws3}
}
}
