\docType{package}
\name{ALA4R}
\alias{ALA4R}
\alias{ALA4R-package}
\title{\pkg{ALA4R}}
\description{
This is a project to enable the R community to access data
and resources hosted at Atlas of Living Australia (ALA).
The goal is to enable outputs (e.g., observations of
species) to be queried and output in standard formats.
}

