% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALA4R.R
\docType{package}
\name{ALA4R}
\alias{ALA4R}
\title{\pkg{ALA4R}
ALA4R}
\description{
`ALA4R` is deprecated and will be removed from CRAN at the end of 2021.  The package to replace `ALA4R``, 
`galah`, is now available on CRAN \url{https://CRAN.R-project.org/package=galah}.
`galah` provides an improved interface to ALA data, while providing the same
core functionality as ALA4R. For an introduction to `galah`, visit the
GitHub page \url{https://github.com/AtlasOfLivingAustralia/galah}.
}
\details{
This project enables the R community to access data and tools hosted by the
Atlas of Living Australia. The goal of the project is
to enable basic species and related information to be queried and output in
standard formats for R. ALA4R is based around the extensive web services
provided by the Atlas; see the API link below.
}
\references{
\url{https://api.ala.org.au/}
}
