% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_occurrences.R
\name{ala_columns}
\alias{ala_columns}
\title{Build dataframe of columns to keep}
\usage{
ala_columns(cols)
}
\arguments{
\item{cols}{string: vector of column names to include. If a col matches a
group name, columns in that group will be added to the list}
}
\description{
Build dataframe of columns to keep
}
