% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_counts.R
\name{ala_counts}
\alias{ala_counts}
\title{Counts for ALA records}
\usage{
ala_counts(taxon_id, filters, geometry, breakdown, limit = 20, caching = FALSE)
}
\arguments{
\item{taxon_id}{string: single species ID or vector of species ids. Use
`ala_taxa()` to lookup species id.}

\item{filters}{data.frame: generated by `ala_filters()`}

\item{geometry}{string: WKT generated by `ala_geometry()`}

\item{breakdown}{field to breakdown the counts by}

\item{limit}{numeric: maximum number of categories to return. 20 by default.
If limit is NULL, all results are returned. For some categories this will
take a while.}

\item{caching}{logical: should the results be cached/use cache? Only used if
`breakdown` is supplied.}
}
\value{
either single integer of total counts, or a dataframe of counts by
`breakdown` field, if specified.
}
\description{
Takes filters in the same format as `ala_occurrences`, with an
additional break down by `breakdown`
}
\examples{
# Count of records in the ALA
ala_counts()
# Counts by state and territory
ala_counts(breakdown = "state")
}
