% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_fields.R
\name{ala_fields}
\alias{ala_fields}
\title{Retrieves a list of all field names that can be used with data
retrieval functions}
\usage{
ala_fields(class = "all")
}
\arguments{
\item{class}{string: class of fields to return e.g. "Assertion"}
}
\value{
data.frame of all fields.
}
\description{
The field names are in Darwin Core format, apart from the fields
which have no DwC equivalent, such as record assertions.
}
\examples{
\dontrun{
 l <- ala_fields()
}
}
\references{
Relevant ALA web services: \itemize{
\item for occurrence fields: \url{https://api.ala.org.au/#ws72}
\item for assertions \url{https://api.ala.org.au/#ws81}
}
}
\seealso{
\code{\link{ala_layers}} to search for spatial layers
}
