% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_occurrences.R
\name{ala_occurrences}
\alias{ala_occurrences}
\title{Download occurrence data from the ALA}
\usage{
ala_occurrences(
  taxon_id,
  filters,
  area,
  columns = "default",
  email = "ala4r@ala.org.au",
  generate_doi = FALSE,
  email_notify = FALSE,
  caching = "off"
)
}
\arguments{
\item{taxon_id}{string: single species ID or vector of species ids. Use
`ala_taxa()` to get lookup species id.}

\item{filters}{data.frame: generated by `ala_filters()`}

\item{area}{string or sf object: restrict the search to an area. Can provide
sf object, or a wkt string. WKT strings longer than 10000 characters will
not be accepted by the ALA- see the vignette for how to work around this.}

\item{columns}{string: vector of columns to return in download.}

\item{email}{string: the email address of the user performing the download
(required unless \code{record_count_only = TRUE}}

\item{generate_doi}{logical: by default no DOI will be generated. Set to
true if you intend to use the data in a publication or similar. If
generated, DOI is stored in attributes.}

\item{email_notify}{logical: set to `FALSE` by default, set to true if you
would like an email notification for the download}

\item{caching}{string: should the results be cached? Either "on" or "off"}
}
\description{
Download occurrence data from the ALA
}
\examples{
\dontrun{
## Retrieve all machine-observed reptile records in Victoria in the past
## five years, with the default ALA data quality profile
id <- ala_taxa("Reptilia")$taxon_concept_id
occ <- ala_occurrences(taxon_id = id, filters =
ala_filters(list(year = seq(2015, 2020), state = "Victoria")))
}
}
