% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_occurrences.R
\name{ala_occurrences}
\alias{ala_occurrences}
\title{Download occurrence data from the ALA}
\usage{
ala_occurrences(
  taxon_id,
  filters,
  area,
  columns = "default",
  email = "ala4r@ala.org.au",
  generate_doi = FALSE,
  email_notify = FALSE
)
}
\arguments{
\item{taxon_id}{string: single species ID or vector of species ids. Use 
`ala_taxa()` to get lookup species id.}

\item{filters}{string: a list to narrow down the search, in the form
`list(field = value)`. To limit the query to a year range, use
`list(year = c(1990, 2000))`. This will return records between and including
the years provided.}

\item{area}{string or sf object: restrict the search to an area. Can provide
sf object, or a wkt string. WKT strings longer than 10000 characters will
not be accepted by the ALA- see the vignette for how to work around this.}

\item{columns}{string: vector of columns to return in download.}

\item{email}{string: the email address of the user performing the download
(required unless \code{record_count_only = TRUE}}

\item{generate_doi}{logical: by default no DOI will be generated. Set to
true if you intend to use the data in a publication or similar}

\item{email_notify}{logical: set to `FALSE` by default, set to true if you
would like an email notification for the download}
}
\description{
Download occurrence data from the ALA
}
