% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_occurrences.R
\name{ala_occurrences}
\alias{ala_occurrences}
\title{Download occurrence data from the ALA.}
\usage{
ala_occurrences(taxon_id, filters, geometry, columns, mint_doi = FALSE)
}
\arguments{
\item{taxon_id}{string: single species ID or vector of species ids. Use
`ala_taxa()` to get lookup species id.}

\item{filters}{data.frame: generated by `ala_filters()`}

\item{geometry}{string: WKT generated by `ala_geometry()`}

\item{columns}{data.frame: generated by `ala_columns()`}

\item{mint_doi}{logical: by default no DOI will be generated. Set to
true if you intend to use the data in a publication or similar}
}
\value{
data.frame of occurrences
}
\description{
If `verbose = TRUE` is set in `ala_config()`, a progress bar will be
displayed for the download.
}
\examples{
\dontrun{
## Retrieve all machine-observed reptile records in Victoria in the past
## five years, with the default ALA data quality profile
id <- ala_taxa("Reptilia")$taxon_concept_id
occ <- ala_occurrences(taxon_id = id, filters =
ala_filters(list(year = seq(2015, 2020), state = "Victoria")))
}
}
