% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_data_profiles.R
\name{ala_quality_filters}
\alias{ala_quality_filters}
\title{Get data filters for a  data quality profile}
\usage{
ala_quality_filters(profile)
}
\arguments{
\item{profile}{string: a data quality profile name, short name or id.
See `ala_data_profiles` for valid filters}
}
\description{
Get data filters for a  data quality profile
}
