% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_taxa.R
\name{ala_taxa}
\alias{ala_taxa}
\title{Function to lookup species information from ALA, given names or unique
identifiers.}
\usage{
ala_taxa(
  term,
  term_type = "name",
  return_children = FALSE,
  include_counts = FALSE
)
}
\arguments{
\item{term}{string, named list or dataframe: search term(s)}

\item{term_type}{string: specifies which type of terms are provided in
`term`. One of name `c('name', 'identifier')`}

\item{return_children}{logical: return child concepts for the provided term(s)}

\item{include_counts}{logical: return occurrence counts for all species returned}
}
\description{
Uses ALA name matching service. Species ids from this function can be
used to search `ala_occurrences`
}
