% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus.R
\docType{data}
\name{aus}
\alias{aus}
\title{Basic Australia raster for visualization
A basic raster dataset in SpatialGridDataFrame (package sp) representing
Australia at 0.05 degree (~5km) resolution.}
\format{
SpatialGridDataFrame (package sp)
}
\usage{
data(aus)
}
\description{
Basic Australia raster for visualization
A basic raster dataset in SpatialGridDataFrame (package sp) representing
Australia at 0.05 degree (~5km) resolution.
}
\keyword{datasets}
