% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\name{data_resources}
\alias{data_resources}
\title{Get a list of current data resources}
\usage{
data_resources(druid, verbose = ala_config()$verbose, max = 100, extra)
}
\arguments{
\item{druid}{string: data resource UID of the data resource(s)}

\item{verbose}{logical: show additional progress information?
[default is set by ala_config()]}

\item{max}{integer: (optional) if all data resources are requested, max
number to return, sorted by record count. Default is top 100}

\item{extra}{string: (optional) additional field to retrieve information for
the data resource. Must be a valid indexed field (see
\code{ala_fields("occurrence_indexed")}}
}
\value{
data frame of data resources
}
\description{
Retrieve a list of all existing data resources, and basic information
for each data resource.
}
\examples{
\dontrun{
# Retrieve information for a dataset
dr_info <- data_resources('dr375')

# Retrieve stats for top 10 data resources with assertions breakdown
dr_info <- data_resources(max = 10, extra = 'assertions')
}
}
\references{
\itemize{
\item Associated ALA web service for listing data resources:
\url{https://collections.ala.org.au/ws/dataResource}
}
}
