\name{species_info}
\alias{species_info}
\title{Taxon or species info}
\usage{
species_info(taxon)
}
\arguments{
  \item{taxon}{a character string for the taxon of
  interest}
}
\value{
a dataframe of information with columns being: \item{comp1
}{Description of 'comp1'}
}
\description{
Returns the profile information for a given taxon. more
needed!!!
}
\details{
NEED
}
\examples{
#find information ALA holds on red kangaroo
	species_info("red kangaroo")
}
\author{
Jeremy VanDerWal \email{jjvanderwal@gmail.com}, Ben Raymond
\email{ben@theraymonds.org}
}
\references{
\url{http://www.ala.org.au/about-the-atlas/downloadable-tools/web-services/}
}

